Fix: useAuth flow
인증 실패시 리다이렉트를 진행해도 이후의 로직이 모두 실행되므로, boolean 값을 반환하도록하여, 인증 실패시 바로 return 처리를 하도록 변경

+++ src/controllers/use-card-editor.ts
@@ -30,7 +30,7 @@ export const useCardEditor = (card: Card) => {
             fileInputRef && fileInputRef.current.click()
         },
         profileChange: async (e: ChangeEvent<HTMLInputElement>) => {
-            await checkAuth()
+            if (!(await checkAuth())) return
             if (!e.target.files || e.target.files.length === 0) return
             setFileLoading(true)
 
@@ -124,7 +124,7 @@ export const useCardEditor = (card: Card) => {
         },
 
         cardDelete: async () => {
-            await checkAuth()
+            if (!(await checkAuth())) return
             publish(PUBSUB_EVENT_TYPES.HIDE_PROFILE_DETAIL)
             if (fileLoading) return
             cardService.delete(card.id).then((res) => {
